Basic things done

+++ app/Http/Controllers/RepliesController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Reply;
 use App\Thread;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Session;
 
 class RepliesController extends Controller
 {
@@ -21,12 +23,29 @@ class RepliesController extends Controller
             'body' => request('body'),
         ]);
 
+        Session::flash('success', 'You have successfully added a reply!');
+
+        $thread->increment('replies');
+
         return back();
     }
 
-    public function destroy(Reply $reply)
+    public function update(Reply $reply)
     {
+        $reply->update(['body' => request('body')]);
+    }
+
+    public function destroy(Reply $reply, Thread $thread)
+    {
+        $this->authorize('delete', $reply);
+
         $reply->delete();
+
+        $reply->thread->decrement('replies');
+
+        if (request()->expectsJson()) {
+            return response(['status' => 'Deleted']);
+        }
 
         return back();
     }
