+++ app/Http/Controllers/ThreadsController.php
@@ -2,12 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use App\User;
+use App\Reply;
 use App\Thread;
 use App\Category;
 use App\Filter\ThreadFilter;
 use Illuminate\Http\Request;
-use Illuminate\Foundation\Auth\User;
+use App\Http\Requests\ThreadRequest;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Session;
 
 class ThreadsController extends Controller
 {
@@ -28,6 +31,8 @@ class ThreadsController extends Controller
 
     public function show($categoryId, Thread $thread)
     {
+        $thread->recordVisit();
+
         return view('show', [
             'thread' => $thread,
         ]);
@@ -40,26 +45,22 @@ class ThreadsController extends Controller
         ]);
     }
 
-    public function store(Request $request)
+    public function store(ThreadRequest $request)
     {
-        $request->validate([
-            'title' => 'required',
-            'body' => 'required',
-        ]);
+        Thread::create($request->validated());
 
-        Thread::create([
-            'user_id' => auth()->id(),
-            'category_id' => $request->category_id,
-            'title' => $request->title,
-            'body' => $request->body,
-        ]);
+        Session::flash('success', 'You have successfully created a thread!');
 
-        return redirect(route('threads.index'));
+        return redirect()->route('threads.index');
     }
 
     public function destroy(Thread $thread)
     {
+        $this->authorize('delete', $thread);
+
         $thread->delete();
+
+        Session::flash('success', 'Your thread has been successfully deleted!');
 
         return redirect()->route('threads.index');
     }
